Changed complexity factor datatype on float

+++ src/ProcessGraph/GraphCalculator.h
@@ -15,7 +15,7 @@ public:
     std::vector<int> GetLateEventDate() { return m_T_late; };
     std::vector<int> GetTimeReserve() { return m_Time_reserve; };
     std::vector<Node*> GetCritPath() { return m_crit_path; };
-    double GetComplexityFactor() { return m_complexity_factor; }
+    float GetComplexityFactor() { return m_complexity_factor; }
 
     // calculation methods
     void CalculateEarlyEventDate();
@@ -31,7 +31,7 @@ private:
     std::vector<int> m_T_early; // EARLY EVENT DATE
     std::vector<int> m_T_late; // LATE EVENT DATE
     std::vector<int> m_Time_reserve; // EVENT TIME RESERVE
-    double m_complexity_factor = 0;
+    float m_complexity_factor = 0;
     std::vector<Node*> m_crit_path; // stores nodes that create a critical path
